added json parsing and improved socket handling also added logging

+++ components/http_server/include/http_server.h
@@ -2,5 +2,5 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-int http_server_start(int port);
+int http_server_start(uint16_t port);
 void http_server_stop();
